increase pointing precision to 32bit

+++ photon_stream_analysis/look_up_table/structure.py
@@ -5,16 +5,16 @@ head = [
     ('event', np.uint16),
     ('reuse', np.uint16),
 
-    ('azimuth', np.float16),
-    ('zenith', np.float16),
+    ('azimuth', np.float32),
+    ('zenith', np.float32),
 
     ('number_photons', np.float16),
     ('cog_cx_pap', np.float16),
     ('cog_cy_pap', np.float16),
 
     ('energy', np.float16),
-    ('theta', np.float16),
-    ('phi', np.float16),
+    ('theta', np.float32),
+    ('phi', np.float32),
     ('impact_x', np.float16),
     ('impact_y', np.float16),
     ('particle', np.uint16),
